+++ client/include/autd3.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 10/09/2021
+// Last Modified: 28/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -33,13 +33,11 @@ using sequence::GAIN_MODE;
 using sequence::GainSequencePtr;
 using sequence::PointSequencePtr;
 
+using core::DEVICE_HEIGHT;
+using core::DEVICE_WIDTH;
 using core::NUM_TRANS_IN_UNIT;
 using core::NUM_TRANS_X;
 using core::NUM_TRANS_Y;
 using core::TRANS_SPACING_MM;
-[[deprecated("please use DEVICE_WIDTH")]] constexpr double AUTD_WIDTH = core::DEVICE_WIDTH;
-[[deprecated("please use DEVICE_HEIGHT")]] constexpr double AUTD_HEIGHT = core::DEVICE_HEIGHT;
-using core::DEVICE_HEIGHT;
-using core::DEVICE_WIDTH;
 using core::ULTRASOUND_FREQUENCY;
 }  // namespace autd